Refactor create a custom hook
Add a custom hook and verify that the tests still pass correctly.
